Moving config_transformers
