+++ recipe/shelf.py
@@ -1,326 +1,28 @@
-import dateparser
-from copy import copy, deepcopy
-
-from datetime import date, datetime
-
-from ordered_set import OrderedSet
+from copy import copy
 from six import iteritems
 from sqlalchemy import (
     Float,
     Integer,
     String,
     Table,
-    and_,
-    case,
-    cast,
-    distinct,
-    func,
-    or_,
 )
-from sqlalchemy.ext.declarative import DeclarativeMeta
-from sqlalchemy.sql.base import ImmutableColumnCollection
 from sqlalchemy.util import lightweight_named_tuple
 from sureberus import errors as E
 from sureberus import normalize_schema
 from yaml import safe_load
 
-from recipe import ingredients
 from recipe.compat import basestring
 from recipe.exceptions import BadIngredient, BadRecipe
 from recipe.ingredients import Dimension, Filter, Ingredient, Metric
-from recipe.schemas import (
-    aggregations,
-    condition_schema,
-    ingredient_schema,
-    shelf_schema,
-    sqlalchemy_datatypes,
-)
-
-# Ensure case and distinct don't get reaped. We need it in scope for
-# creating Metrics
-_distinct = distinct
-_case = case
+from recipe.schemas import shelf_schema
 
 _POP_DEFAULT = object()
 
-# constant used for ensuring safe division
-SAFE_DIVISON_EPSILON = 0.000000001
 
+def ingredient_transformer(ingr):
+    version = ingr.get('_version', '1')
 
-def ingredient_class_for_name(class_name):
-    """Get the class in the recipe.ingredients module with the given name."""
-    return getattr(ingredients, class_name, None)
-
-
-def _find_in_columncollection(columns, name):
-    """ Find a column in a column collection by name or _label"""
-    for col in columns:
-        if col.name == name or getattr(col, "_label", None) == name:
-            return col
-    return None
-
-
-def find_column(selectable, name):
-    """
-    Find a column named `name` in selectable
-
-    :param selectable:
-    :param name:
-    :return: A column object
-    """
-    from recipe import Recipe
-
-    if isinstance(selectable, Recipe):
-        selectable = selectable.subquery()
-
-    # Selectable is a table
-    if isinstance(selectable, DeclarativeMeta):
-        col = getattr(selectable, name, None)
-        if col is not None:
-            return col
-
-        col = _find_in_columncollection(selectable.__table__.columns, name)
-        if col is not None:
-            return col
-
-    # Selectable is a sqlalchemy subquery
-    elif hasattr(selectable, "c") and isinstance(
-        selectable.c, ImmutableColumnCollection
-    ):
-        col = getattr(selectable.c, name, None)
-        if col is not None:
-            return col
-
-        col = _find_in_columncollection(selectable.c, name)
-        if col is not None:
-            return col
-
-    raise BadIngredient("Can not find {} in {}".format(name, selectable))
-
-
-def _convert_date_value(v):
-    parse_kwargs = {"languages": ["en"]}
-    if isinstance(v, date):
-        return v
-    elif isinstance(v, datetime):
-        return v.date()
-    elif isinstance(v, basestring):
-        parsed_dt = dateparser.parse(v, **parse_kwargs)
-        if parsed_dt is None:
-            raise ValueError("Could not parse date in {}".format(v))
-        return parsed_dt.date()
-    else:
-        raise ValueError("Can not convert {} to date".format(v))
-
-
-def _convert_datetime_value(v):
-    parse_kwargs = {"languages": ["en"]}
-    if isinstance(v, datetime):
-        return v
-    elif isinstance(v, date):
-        return datetime(v.year, v.month, v.day)
-    elif isinstance(v, basestring):
-        parsed_dt = dateparser.parse(v, **parse_kwargs)
-        if parsed_dt is None:
-            raise ValueError("Could not parse datetime in {}".format(v))
-        return parsed_dt
-    else:
-        raise ValueError("Can not convert {} to datetime".format(v))
-
-
-def convert_value(field, value):
-    """Convert values into something appropriate for this SQLAlchemy data type
-
-    :param field: A SQLAlchemy expression
-    :param values: A value or list of values
-    """
-
-    if isinstance(value, (list, tuple)):
-        if str(field.type) == "DATE":
-            return [_convert_date_value(v) for v in value]
-        elif str(field.type) == "DATETIME":
-            return [_convert_datetime_value(v) for v in value]
-        else:
-            return value
-    else:
-        if str(field.type) == "DATE":
-            return _convert_date_value(value)
-        elif str(field.type) == "DATETIME":
-            return _convert_datetime_value(value)
-        else:
-            return value
-
-
-def parse_validated_condition(cond, selectable):
-    """ Convert a validated condition into a SQLAlchemy boolean expression """
-    if cond is None:
-        return
-
-    if "and" in cond:
-        conditions = []
-        for c in cond.get("and", []):
-            conditions.append(parse_validated_condition(c, selectable))
-        return and_(*conditions)
-
-    elif "or" in cond:
-        conditions = []
-        for c in cond.get("or", []):
-            conditions.append(parse_validated_condition(c, selectable))
-        return or_(*conditions)
-
-    elif "field" in cond:
-        field = parse_validated_field(cond.get("field"), selectable)
-        _op = cond.get("_op")
-        _op_value = convert_value(field, cond.get("_op_value"))
-
-        if _op == "between":
-            return getattr(field, _op)(*_op_value)
-        else:
-            return getattr(field, _op)(_op_value)
-
-
-def parse_unvalidated_condition(cond, selectable):
-    if cond is None:
-        return
-    try:
-        cond = normalize_schema(condition_schema, cond, allow_unknown=False)
-    except E.SureError as e:
-        raise BadIngredient(str(e))
-    return parse_validated_condition(cond, selectable)
-
-
-def parse_unvalidated_field(unvalidated_fld, selectable, aggregated=True):
-    kind = "Metric" if aggregated else "Dimension"
-    ingr = {"field": unvalidated_fld, "kind": kind}
-    try:
-        ingr_dict = normalize_schema(ingredient_schema, ingr, allow_unknown=True)
-    except E.SureError as e:
-        raise BadIngredient(str(e))
-    return parse_validated_field(ingr_dict["field"], selectable)
-
-
-def ingredient_from_unvalidated_dict(unvalidated_ingr, selectable):
-    try:
-        ingr_dict = normalize_schema(
-            ingredient_schema, unvalidated_ingr, allow_unknown=True
-        )
-    except E.SureError as e:
-        raise BadIngredient(str(e))
-    return ingredient_from_validated_dict(ingr_dict, selectable)
-
-
-def parse_validated_field(fld, selectable, use_bucket_labels=True):
-    """ Converts a validated field to a sqlalchemy expression.
-    Field references are looked up in selectable """
-    if fld is None:
-        return
-
-    fld = deepcopy(fld)
-
-    if fld.pop("_use_raw_value", False):
-        return float(fld["value"])
-
-    if "buckets" in fld:
-        # Buckets only appear in dimensions
-        buckets_default_label = (
-            fld.get("buckets_default_label") if use_bucket_labels else 9999
-        )
-        conditions = [
-            (
-                parse_validated_condition(cond, selectable),
-                cond.get("label") if use_bucket_labels else idx,
-            )
-            for idx, cond in enumerate(fld.get("buckets", []))
-        ]
-        field = case(conditions, else_=buckets_default_label)
-    else:
-        field = find_column(selectable, fld["value"])
-
-    operator_lookup = {
-        "+": lambda fld: getattr(fld, "__add__"),
-        "-": lambda fld: getattr(fld, "__sub__"),
-        "/": lambda fld: getattr(fld, "__div__"),
-        "*": lambda fld: getattr(fld, "__mul__"),
-    }
-    for operator in fld.get("operators", []):
-        op = operator["operator"]
-        other_field = parse_validated_field(operator["field"], selectable)
-        if op == "/":
-            other_field = (
-                func.coalesce(cast(other_field, Float), 0.0) + SAFE_DIVISON_EPSILON
-            )
-        field = operator_lookup[op](field)(other_field)
-
-    # Apply a condition if it exists
-    cond = parse_validated_condition(fld.get("condition", None), selectable)
-    if cond is not None:
-        field = case([(cond, field)])
-
-    # Lookup the aggregation function
-    aggr_fn = aggregations.get(fld.get("aggregation"))
-    field = aggr_fn(field)
-
-    # lookup the sqlalchemy_datatypes
-    cast_to_datatype = sqlalchemy_datatypes.get(fld.get("_cast_to_datatype"))
-    if cast_to_datatype is not None:
-        field = cast(field, cast_to_datatype)
-
-    coalesce_to_value = fld.get("_coalesce_to_value")
-    if coalesce_to_value is not None:
-        field = func.coalesce(field, coalesce_to_value)
-
-    return field
-
-
-def ingredient_from_validated_dict(ingr_dict, selectable):
-    """ Create an ingredient from an dictionary.
-
-    This object will be deserialized from yaml """
-
-    kind = ingr_dict.pop("kind", "Metric")
-    IngredientClass = ingredient_class_for_name(kind)
-
-    if IngredientClass is None:
-        raise BadIngredient("Unknown ingredient kind")
-
-    field_defn = ingr_dict.pop("field", None)
-    divide_by_defn = ingr_dict.pop("divide_by", None)
-
-    field = parse_validated_field(field_defn, selectable, use_bucket_labels=True)
-    if isinstance(field_defn, dict) and "buckets" in field_defn:
-        ingr_dict["order_by_expression"] = parse_validated_field(
-            field_defn, selectable, use_bucket_labels=False
-        )
-
-    if divide_by_defn is not None:
-        # Perform a divide by zero safe division
-        divide_by = parse_validated_field(divide_by_defn, selectable)
-        field = cast(field, Float) / (
-            func.coalesce(cast(divide_by, Float), 0.0) + SAFE_DIVISON_EPSILON
-        )
-
-    quickselects = ingr_dict.pop("quickselects", None)
-    parsed_quickselects = []
-    if quickselects:
-        for qf in quickselects:
-            parsed_quickselects.append(
-                {
-                    "name": qf["name"],
-                    "condition": parse_validated_condition(
-                        qf.get("condition", None), selectable
-                    ),
-                }
-            )
-    ingr_dict["quickselects"] = parsed_quickselects
-
-    args = [field]
-    # Each extra field contains a name and a field
-    for extra in ingr_dict.pop("extra_fields", []):
-        ingr_dict[extra.get("name")] = parse_validated_field(
-            extra.get("field"), selectable
-        )
-
-    return IngredientClass(*args, **ingr_dict)
+from recipe.schemas.config_transformers import  ingredient_from_validated_dict, parse_unvalidated_condition
 
 
 class Shelf(object):
